Moved validateConfig to lib and added portal config rules
